Consolidated CDC posts for pregnacy and events
Updated: I am pregnant and worried about COVID-19. What are the risks to me and my baby?
Deleted What is the risk to pregnant women of getting COVID-19? and redirected to "I am pregnant."
Deleted Can COVID-19 be passed from mother to child? and redirected to "I am pregnant."
Updated Do I need to cancel an event if there are confirmed cases of COVID-19 in the community?
Deleted When does the CDC recommend that I cancel or postpone an event? and redirected to "Do I need to cancel"
Updated the category ordering to reflect these changes.